Alter history of updates window sizing
Set the defaults so that it opens with all columns visible and a size that looks about right in relation to the main window.  Set the minimum width so that it can be trimmed down to show just date and package, and the height can be reduced a little too if required.