Hide mintUpdate status icon in certain cases
Currently, mintUpdate's status icon is always shown. I think this is unnecessary in certain situations.
This makes the icon invisible if the system is up to date, updates are being searched, APT is used by something else or the sources couldn't be reloaded, since I don't think it's necessary to show the icon in this case.
The icon becomes visible when new updates are found, being installed, or any grave errors occur.