Use indicator applet
As most Cinnamon applets use indicators instead of system trays and
indicators are automatically themed when using certain icons, Mint
Update was updated to use an Indicator.

Sadly, the underlying indicator library can only show a menu on click,
clicks on the icon are not customizable, and there is no signal for
double-click. Hence, in order to show the main window, one needs to
click on a menu item now.
Also there is no tooltip support (who came up with such a library
anyway?) so status information will be displayed inside the menu.

For automatic theming, icons need to be suffixed with -symbolic.svg.
See my next commit.

This closes linuxmint/mintupdate#101