mintUpdate.py: don't download changelog each time on tab switch
Fixes #26

While viewing the same package, each time a switch was done from the Description to the Changelog tab the changelogs would be downloaded anew.

The function switch_page is called when switching between tabs while viewing the same package. This function is changed to not erase the contents of the Description and Changelog tabs. The changelogs will only be downloaded if the Changelog tab is being switched to and haven't been downloaded yet while viewing this package.

The function display_selected_package is called when selected another package to view. This function is changed to always set the Description tab.