Fix and improve kernel updates
Kernel updates were broken in previous commits, by the addition of the source column.
This commit fixes it.

Kernel updates previously mixed updates for Mint's recommended kernel and Ubnutu's recommended kernel.
This is a rare case and it only happens if the user upgraded the system himself and is running an old kernel.
This commit fixes that situation and shows the two kernels as separate updates.

Kernel updates were distinct from linux-libc-dev updates. They're now mixed with it if the versions match.

The logic used to identify an update as a package update, a security update or a kernel update was refined.
In the case of kernel updates, the emphasis is on the fact that they are kernel updates, whether or not they're security updates.
The rationale for this is that that type of update is configurable and it is distinct from the security updates configuration.

Aliases were refined for kernel updates to clearly show which version of the kernel is at play and emphasize the idea that
kernel updates aren't really "updates", but the arrival of a "new" kernel.