Add CLI support
This commit introduces a new command-line utility called mintupdate-tool.

That utility can list and apply available updates.

It uses the mintupdate settings to decide how to filter updates, but also
accept arguments which can override them.

A lot of code was removed from mintupdate.py and placed in Classes.py.

Many features were moved from mintupdate.py (which is the UI part of mintupdate)
to checkAPT.py (which is used by both the UI and the CLI).

mintupdate-tool can be used in a terminal or a script or even a cron job
to automate updates.