Lower Firefox to a level 1 update
It is multi-purpose and central to the desktop and its rapid pace of development makes it
prone to regressions and one of the most sensitive applications out there.

Its level can however be lowered for the following reasons:

- In line with the definition of a level 1 update, it doesn't have any impact other than on itself.
- It's a very well known application. Users are quick to identify the update in case of regressions.

Although the lack of a functional web browser could be a huge issue for novice users, the security nature of all Firefox
updates also makes this update a mandatory one. The role of the level here is therefore limited to raising awareness, and this
is not needed for such a well identified update.