Remove level 5 from UI/configuration
Levels describe the impact an update has on the system and range from a value of 1 to 4.

The mechanism in place to temporarily block an update is to set its level to 5. This mechanism
was introduced in 18.3 and was never used (we usually either fix the problem directly, or pin Ubuntu with
a patch, so in theory we shouldn't even need this mechanism, but it's there anyway to empower us to block
an update if we find a critical regression and no solution is available yet).

Although this mechanism uses level=5 from a technical point of view, to the user, this has nothing to
do with levels at all. Whereas a user should ALWAYS apply updates no matter what levels they are, if we were
ever to decide to block an update, users should NEVER apply it (for as long as it's blocked).

With this in mind, it's important not to expose users to blocked updates and not to confuse levels with this
mechanism.