Replace update policies with a welcome screen.
The introduction of system snapshots allows us to
recommend all updates without caution.

The concept of regression is still important to explain
but the recommendation becomes completely different.

Rather than to recommend caution (although that will
still be important when troubleshooting a regression),
we can recommend system snapshots.

The core concepts remain the same (stability-regression and
security/bug-fixes), but no longer have to be balanced against
one another.

The message is much clearer to the user and doesn't lead to misinterpretation:

- Set up snapshots
- Update everything