CheckAPT: Improve/Fix kernel updates
First, this fixes kernel updates for 18.04 and 4.15+ in 16.04,
where the kernel series is more recent than the meta,
and the naming convention has changed (i.e. moved from linux-image-extra to
linux-modules).

Second, this commit refines kernel updates. It only provides one meta (this
was only an issue in 18 ALPHA as 16.04-hwe had the same version as the generic
meta), and only does so if no metas are installed.