Add support for lowlatency kernels
Don't expose this in the preferences, "lowlatency" sounds cool,
we don't want novice users to switch to that without knowing what
it is.

The assumption is that lowlatency kernel users are advanced to start
with and will therefore use gsettings to switch mintupdate from supporting
generic series to lowlatency ones.