Run automatic updates in non-interactive mode
When running automatic updates, this prevents debconf to ask things to the user
and to wait endlessly (because it is not possible for the user to answer).

When this happened, some package were left unconfigured and user had to open a
terminal to type "sudo dpkg --configure -a" to be able to install/update
software again.