Present the window if the welcome screen needs to be acknowledged
The Update Manager starts to refresh the cache and look for updates by itself (auto-refresh mode
is enabled by default) only after the welcome screen is acknowledged by the user.

If the welcome screen is still not acknowleged, this commit maps the window if it isn't visible before
the refresh.

On a fresh installation this means the update manager window will popup by itself after 10 minutes,
and then every 2 hours until the user presses OK.