Avoid unreachable message for non-HTTP mirrors (#701)
When using a scheme for a mirror that is not HTTP, such as ftp:,
mirror:, or file:, mintupdate reports a message that the mirror cannot
be reached. In actuality, the mirror very well may be reachable but it
cannot check that using an HTTP request. Rather than trying to implement
support for all the protocols in sources.list(5), this skips the check
for any mirror_url that doesn't start with `http`.


ref: https://github.com/jmunixusers/cs-vm-build/pull/475