Make sure to always use the proper flatpak symbolic icon.
The existing icon was added to the icon theme at runtime via
gtk_icon_theme_append_path. Unfortunately custom paths are the
very last thing looked at when the theme engine looks for an icon,
choosing to go with 'flatpak' first if the current theme has it,
instead of the exact match that we added.