mintupdate: Use a C long type for the update size column.
Python ints are turned into longs when they overflow, but when
defining a GtkTreeStore, the int 'type' is strictly interpreted,
causing an overflow error if the package's update size is more
than 2gb.

Since this is now possible with flatpak updates, use a real long
type when defining the store.

Fixes #781