Maintain a useful height for the details pane when using a low- resolution monitor.
The details pane was shrinking to accomodate lower resolutions,
to the point it was pretty useless without having to resize.

This will keep the lower pane at a consistent size, and adjustments
for monitor height will be taken from the upper pane instead
where there is a lot more spare room. Setting the 'shrink' property
to true makes sure the lower pane can still be made smaller if
the user wants to expand the update list.

Fixes #831