Build the update list while still in the refresh thread.
The refresh thread was getting a list of APT updates and then
calling refresh_updates() on an idle callback. This would cause
the UI to hang while getting cinnamon and flatpak updates.

Remove the _idle wrapper from refresh_updates() and build a simple
array of Update objects info to pass to show_updates_in_UI (which
is an idle callback).