mintUpdate.py: Show window as busy immediately upon starting an update operation.
There was a delay between starting an install and blocking the ui
to prevent further (undefined) interaction that could allow
multiple dialogs to appear.

Remove the idle delay from set_window_busy(), so it applies
immediately when called, but keep it on an idle callback in
finish_install(), which is not on the main thread.

Fixes https://github.com/linuxmint/mint22.2-beta/issues/34.